hand experiment update - dump files from logs + create PDMs from files
