+++ src/main/scala/apps/femur/CompareModels.scala
@@ -1,3 +1,19 @@
+/*
+ *  Copyright University of Basel, Graphics and Vision Research Group
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
 package apps.femur
 
 import java.io._
@@ -48,8 +64,7 @@ object CompareModels {
     println("Computing specificity")
     (1 to model.rank).map { i =>
       val pdm = PointDistributionModel(model.truncate(i).gp)
-      val tmp = ModelMetrics.specificity(pdm, trainingData, 1000)
-      tmp
+      ModelMetrics.specificity(pdm, trainingData, 1000)
     }
   }
 
@@ -58,8 +73,7 @@ object CompareModels {
     val dc: TriangleMeshDataCollection[_3D] = DataCollection.fromTriangleMesh3DSequence(model.referenceMesh, meshes)
     val res = (1 to model.rank).map { i =>
       val pdm = PointDistributionModel(model.truncate(i).gp)
-      val tmp = ModelMetrics.generalization(pdm, dc).get
-      tmp
+      ModelMetrics.generalization(pdm, dc).get
     }
     res
   }
@@ -67,8 +81,7 @@ object CompareModels {
   def computeModelMetrics(): ListBuffer[jsonModelComparisonFormatClass] = {
     val logger: ListBuffer[jsonModelComparisonFormatClass] = new ListBuffer[jsonModelComparisonFormatClass]
 
-    val meshesGtPath = new File(basePath, "registered/meshes").listFiles(f => f.getName.endsWith(".stl") && f.getName.replace(".stl", "").toInt < modelIndexThreshold)
-    val meshesValidationPath = new File(basePath, "registered/meshes").listFiles(f => f.getName.endsWith(".stl") && f.getName.replace(".stl", "").toInt >= modelIndexThreshold)
+    val meshesGtPath = new File(basePath, "registered/meshes").listFiles(f => f.getName.endsWith(".vtk") && f.getName.replace(".vtk", "").toInt < modelIndexThreshold)
 
     val gtMeshes = meshesGtPath.map(f => MeshIO.readMesh(f).get)
     val validationMeshes = meshesGtPath.map(f => MeshIO.readMesh(f).get)
@@ -81,7 +94,7 @@ object CompareModels {
     val meanModel = StatisticalModelIO.readStatisticalMeshModel(meanModelFile).get
     val samplesModelsPath = samplePath.listFiles(_.getName.endsWith(".h5"))
 
-    println("GT model")
+    println("Model from complete data")
     logger += jsonModelComparisonFormatClass(
       "gt",
       gtModelFile.getName.replace(".h5", ""),
