+++ src/main/scala/apps/hands/DumpMeshesFromLog.scala
@@ -17,12 +17,13 @@
 package apps.hands
 
 import java.io.File
+
 import apps.util.LogHelper2D
 import scalismo.io.{MeshIO, StatisticalModelIO}
 
 import scala.collection.parallel.ForkJoinTaskSupport
 
-object Logs2Meshes {
+object DumpMeshesFromLog {
 
 
   def main(args: Array[String]) {
@@ -33,22 +34,26 @@ object Logs2Meshes {
     val modelFile = new File(Paths.handPath, "hand2D_gp_s25_s50_s120_per.h5")
     val model = StatisticalModelIO.readStatisticalLineMeshModel2D(modelFile).get
 
-    val experimentPath = new File(Paths.handPath, s"experiments/shapemi/allFingers15pMissing")
+    val experimentPath = new File(Paths.handPath, s"shapemi/allFingers15pMissing")
     experimentPath.mkdirs()
 
     println(s"ExperimentPATH: ${experimentPath}")
+    val randomPath = new File(experimentPath, s"samples/")
+    val meanPath = new File(experimentPath, s"mean/")
+    val mapPath = new File(experimentPath, s"map/")
+    val outPaths = Seq(randomPath, meanPath, mapPath)
+    outPaths.foreach(_.mkdir())
 
     val percentageCuts = Seq(15)
     val fingerSeq = Seq("thumb", "index", "long", "ring", "small")
 
     val idSeq = 0 to 11
-    val targetNames = idSeq.map(i => s"hand-${i}_${fingerSeq(i%fingerSeq.length)}_${percentageCuts(i%percentageCuts.length)}").par
+    println("Targets: ")
+    val targetNames = idSeq.map(i => s"hand-${i}_${fingerSeq(i % fingerSeq.length)}_${percentageCuts(i % percentageCuts.length)}").sorted
 
-    targetNames.foreach(println(_))
+    targetNames.foreach(f => println(s" - ${f}"))
 
-    targetNames.tasksupport = new ForkJoinTaskSupport(new java.util.concurrent.ForkJoinPool(10))
-
-    targetNames.zipWithIndex.foreach { case (targetpartialname, sleepTime) =>
+    targetNames.zipWithIndex.par.foreach { case (targetpartialname, sleepTime) =>
 
       Thread.sleep(sleepTime * 100)
 
@@ -61,19 +66,12 @@ object Logs2Meshes {
       val best = log.mapMesh2D()
       val mean = log.meanMesh2D(takeEveryN = 50)
 
-      println(s"Output folders will be created under: ${experimentPath}")
-      val randomPath = new File(experimentPath, s"samples/")
-      val meanPath = new File(experimentPath, s"mean/")
-      val mapPath = new File(experimentPath, s"map/")
-      val outPaths = Seq(randomPath, meanPath, mapPath)
-      outPaths.foreach(_.mkdir())
-
       MeshIO.writeLineMesh(mean, new File(meanPath, targetpartialname + "_mean.vtk"))
-      MeshIO.writeLineMesh(best, new File(mapPath, targetpartialname + "_map.vtk"))
+//      MeshIO.writeLineMesh(best, new File(mapPath, targetpartialname + "_map.vtk"))
 
-      rndShapes.take(50).zipWithIndex.foreach { case (m, i) =>
-        MeshIO.writeLineMesh(m, new File(randomPath, targetpartialname + s"_${i}.vtk"))
-      }
+//      rndShapes.take(10).zipWithIndex.foreach { case (m, i) =>
+//        MeshIO.writeLineMesh(m, new File(randomPath, targetpartialname + s"_${i}.vtk"))
+//      }
     }
   }
 }
