+++ src/main/scala/apps/hands/RegistrationAll.scala
@@ -20,42 +20,41 @@ import java.io.File
 
 import scalismo.geometry._
 import scalismo.io.{LandmarkIO, StatisticalModelIO}
-import scalismo.ui.api.{ScalismoUI, ScalismoUIHeadless}
+import scalismo.ui.api.ScalismoUIHeadless
 
-object RegistrationAll extends App{
+object RegistrationAll extends App {
 
-    println("starting app...")
-    scalismo.initialize()
+  println("starting app...")
+  scalismo.initialize()
 
-    val modelFile = new File(Paths.handPath, "hand2D_gp_s25_s50_s120_per.h5")
-    val model = StatisticalModelIO.readStatisticalLineMeshModel2D(modelFile).get
-    val modelLMFile = new File(Paths.handPath, "reference-hand.json")
-    val modelLM: Seq[Landmark[_2D]] = LandmarkIO.readLandmarksJson2D(modelLMFile).get
+  val modelFile = new File(Paths.handPath, "hand2D_gp_s25_s50_s120_per.h5")
+  val model = StatisticalModelIO.readStatisticalLineMeshModel2D(modelFile).get
+  val modelLMFile = new File(Paths.handPath, "reference-hand.json")
+  val modelLM: Seq[Landmark[_2D]] = LandmarkIO.readLandmarksJson2D(modelLMFile).get
 
-    val targetFiles = new File(Paths.handPath, s"aligned/mesh")
-//    val fingers = Seq("thumb", "index", "long", "ring", "small")
-//    val cutPercentages = Seq(5, 10, 15, 20, 30, 40, 50, 60, 70, 80, 90)
-    val fingers = Seq("thumb")
-    val cutPercentages = Seq(15)
+  val targetFiles = new File(Paths.handPath, s"aligned/mesh")
+  val fingers = Seq("thumb", "index", "long", "ring", "small")
 
-    val logPath = Paths.handLogPath
-    logPath.mkdir()
+  val cutPercentages = Seq(5, 10, 15, 20, 30, 40, 50, 60, 70, 80, 90)
 
-    targetFiles.listFiles(_.getName.endsWith(".vtk")).sorted.par.foreach{targetGTFile =>
-        val targetName = targetGTFile.getName.replace(".vtk", "")
-        fingers.foreach{finger =>
-          cutPercentages.foreach{cutPercentage =>
-              val partialName = s"${targetName}_${finger}_${cutPercentage}"
-              val targetFile = new File(Paths.handPath, s"partial/mesh/${partialName}.vtk")
-              val targetLMFile = new File(Paths.handPath, s"partial/landmarks/${partialName}.json")
+  val logPath = Paths.handLogPath
+  logPath.mkdir()
 
-              val log = new File(logPath, s"${partialName}.json")
+  targetFiles.listFiles(_.getName.endsWith(".vtk")).sorted.par.foreach { targetGTFile =>
+    val targetName = targetGTFile.getName.replace(".vtk", "")
+    fingers.foreach { finger =>
+      cutPercentages.foreach { cutPercentage =>
+        val partialName = s"${targetName}_${finger}_${cutPercentage}"
+        val targetFile = new File(Paths.handPath, s"partial/mesh/${partialName}.vtk")
+        val targetLMFile = new File(Paths.handPath, s"partial/landmarks/${partialName}.json")
 
-              val ui = ScalismoUIHeadless()
+        val log = new File(logPath, s"${partialName}.json")
 
-              val reg = HandRegistration(model, modelLM, targetGTFile, targetFile, targetLMFile, log)
-              reg.run(ui = ui, numOfSamples = 1000)
-          }
-        }
+        val ui = ScalismoUIHeadless()
+
+        val reg = HandRegistration(model, modelLM, targetGTFile, targetFile, targetLMFile, log)
+        reg.run(ui = ui, numOfSamples = 1000)
+      }
     }
+  }
 }
